requirements.txt
pandas==2.2.3
torch==2.5.1
sentence-transformers==3.3.1
pyfiglet==1.0.2
matplotlib==3.10.0